Added spring boot sql queries for filtering cars on rentals-page. Possible filter options: sort-order (price, alphabetical), car-type, transmission, and passenger-capacity.

+++ src/main/java/no/ntnu/entity/models/Cars.java
@@ -61,7 +61,7 @@ public class Cars {
 
   @Column(nullable = false)
   @ApiModelProperty("The number of passengers the car can carry")
-  private byte passengers;
+  private int passengers;
 
   @Column(nullable = false)
   @Enumerated(EnumType.STRING)
@@ -169,7 +169,7 @@ public class Cars {
     this.productionYear = productionYear;
   }
 
-  public byte getPassengers() {
+  public int getPassengers() {
     return passengers;
   }
